requirements.txt
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-pinecone==0.2.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
pinecone-client==5.0.0
google-generativeai==0.7.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3